remove unused errors file. move pagination file
